refactor(settings): SettingsCoordinator の純粋正規化関数を SettingsNormalization へ抽出 (#177)
- NormalizeLanguageSetting / NormalizeMapZoomLevel / NormalizeExternalContentBaseUrl /
  NormalizePaneLayoutPreset / NormalizePaneRegionViews / FindValidAncestorPath を
  Services/SettingsNormalization.cs へ移動し、SettingsCoordinator を
  オーケストレーション責務に縮退（621 行 → 417 行）
- SettingsPaneViewModel の重複正規化ロジックを抽出先へ集約
  （言語は完全版へ委譲、ズームは SnapMapZoomLevelToNearest として意図を分離）
- ISettingsCoordinator は全メンバー使用中のためスリム化対象なし
- 正規化テストを SettingsNormalizationTests へ移動し境界値を追加（計 51 ケース）

Refs #150
