+++ PhotoGeoExplorer.Tests/SettingsCoordinatorTests.cs
@@ -180,103 +180,6 @@ public sealed class SettingsCoordinatorTests : IDisposable
         Assert.Equal(1, eventCount);
     }
 
-    [Theory]
-    [InlineData(null, null)]
-    [InlineData("", null)]
-    [InlineData("system", null)]
-    [InlineData("ja", "ja-JP")]
-    [InlineData("ja-JP", "ja-JP")]
-    [InlineData("en", "en-US")]
-    [InlineData("en-us", "en-US")]
-    [InlineData("fr-FR", "fr-FR")]
-    public void NormalizeLanguageSettingReturnsExpectedValue(string? input, string? expected)
-    {
-        var actual = SettingsCoordinator.NormalizeLanguageSetting(input);
-        Assert.Equal(expected, actual);
-    }
-
-    [Theory]
-    [InlineData(8, 8)]
-    [InlineData(14, 14)]
-    [InlineData(18, 18)]
-    [InlineData(7, 14)]
-    [InlineData(20, 14)]
-    public void NormalizeMapZoomLevelReturnsExpectedValue(int input, int expected)
-    {
-        var actual = SettingsCoordinator.NormalizeMapZoomLevel(input);
-        Assert.Equal(expected, actual);
-    }
-
-    [Theory]
-    [InlineData(null, null)]
-    [InlineData("", null)]
-    [InlineData("https://photogeoexplorer.pages.dev", "https://photogeoexplorer.pages.dev")]
-    [InlineData("https://photogeoexplorer.pages.dev/", "https://photogeoexplorer.pages.dev")]
-    [InlineData("http://example.com/help", "http://example.com/help")]
-    [InlineData("ftp://example.com", null)]
-    public void NormalizeExternalContentBaseUrlReturnsExpectedValue(string? input, string? expected)
-    {
-        var actual = SettingsCoordinator.NormalizeExternalContentBaseUrl(input);
-        Assert.Equal(expected, actual);
-    }
-
-    [Theory]
-    [InlineData(0)]
-    [InlineData(1)]
-    [InlineData(2)]
-    public void NormalizePaneLayoutPresetReturnsInputWhenDefined(int input)
-    {
-        var preset = (PaneLayoutPreset)input;
-        var actual = SettingsCoordinator.NormalizePaneLayoutPreset(preset);
-        Assert.Equal(preset, actual);
-    }
-
-    [Fact]
-    public void NormalizePaneLayoutPresetReturnsDefaultForInvalidValue()
-    {
-        var actual = SettingsCoordinator.NormalizePaneLayoutPreset((PaneLayoutPreset)999);
-        Assert.Equal(AppSettings.DefaultPaneLayoutPreset, actual);
-    }
-
-    [Fact]
-    public void NormalizePaneRegionViewsReturnsSameValuesWhenDistinct()
-    {
-        var actual = SettingsCoordinator.NormalizePaneRegionViews(
-            PaneViewType.Map,
-            PaneViewType.Preview,
-            PaneViewType.File);
-
-        Assert.Equal(PaneViewType.Map, actual.Region1View);
-        Assert.Equal(PaneViewType.Preview, actual.Region2View);
-        Assert.Equal(PaneViewType.File, actual.Region3View);
-    }
-
-    [Fact]
-    public void NormalizePaneRegionViewsReplacesDuplicateWithUnusedValue()
-    {
-        var actual = SettingsCoordinator.NormalizePaneRegionViews(
-            PaneViewType.File,
-            PaneViewType.File,
-            PaneViewType.Map);
-
-        Assert.Equal(PaneViewType.File, actual.Region1View);
-        Assert.Equal(PaneViewType.Preview, actual.Region2View);
-        Assert.Equal(PaneViewType.Map, actual.Region3View);
-    }
-
-    [Fact]
-    public void NormalizePaneRegionViewsNormalizesInvalidAndRemovesDuplicates()
-    {
-        var actual = SettingsCoordinator.NormalizePaneRegionViews(
-            (PaneViewType)99,
-            (PaneViewType)100,
-            PaneViewType.File);
-
-        Assert.Equal(PaneViewType.File, actual.Region1View);
-        Assert.Equal(PaneViewType.Preview, actual.Region2View);
-        Assert.Equal(PaneViewType.Map, actual.Region3View);
-    }
-
     public void Dispose()
     {
         foreach (var dir in _tempDirectories)
